remove Frozen_Start functionality
The main issue is that we cannot enableSimulation just locally,
the state becomes desynchronized, server detect this and forces
sync, enabling simulation.
This is hard to reproduce as it needs 2 people + dedicated server.

We could use enableSimulationGlobal, but that needs to run on the
server, which means either split the logic for client/server or
do remoteExec for disable/enable.

The former is very hard to do due to JIP clients (and the server
not being in init.sqf anymore) and the latter just means a lot more
traffic and overhead for functionality that's supposed to prevent
overhead.

Therefore ditch the feature in its current state.

Signed-off-by: freghar <freghar@dummy.tld>