Revert "add fn_triggerSafeArea for making accident-free zones"
This function was made on a modded game where "ProtectionZone_Ep1"
was present (CUP Terrains). It was only later when I realized this
object isn't in vanilla A3, despite it being in A2.

Given this revelation, there are better ways to do similar thing
in vanilla A3, for example Repeatable trigger doing allowDamage
false on activation (in trigger area) / true on deactivation (outside
trigger area). The bonus is that it works against moving objects
crushing players and has configurable area shape and size.

This reverts commit 4f00e1215084ef964797e36249698d097f7e4ee9.