remove realistic example from Custom_Factions
This is to avoid conflicts with templates/missions based on this one
- these would have to remove our example and, when we update it as
a result of an Arma update, get conflicts in the code.

Many git users are not trained how to deal with conflict resolution,
so try to minimize the risk to a minimum.

The examples given in the README should be enough to make the user
understand the syntax.

Signed-off-by: freghar <freghar@dummy.tld>