Revert "optimize per-unit loadouts in Custom_Factions"
While there was one known downside of the per-unit CODEs not executing
after a locality change (due to their lack on the new client), this was
supposed to not be an issue as player-controlled clients always execute
the init lines on JIP, setting up the per-unit CODEs.

The problem is that there's one additional race between the initial unit
initialization (which happens on the server), the locality change of the
player unit (even if spawned anew) and the executed init line.

In short, just revert to storing the CODEs with the unit - it's simple,
works across any locality changes, across JIPs, respawns, headless
clients, etc. and is actually optimized/preprocessed after the first run
and race-free / reliable.

This reverts commit 5bcdeeb185a513c24b1a6fcd8a5af8ecd5d659cf.