broadcast "A3MT_factionsLoaded" in Custom_Factions
Because

  1) the class-based loadouts get initially triggered on the server
  2) then the (player) unit is transferred to the player client
  3) and after that, the init line is executed by the player client

If we don't broadcast this, the variable won't be seen by the player
client and its waitUntil loop will get stuck.

This also allows us to call fn_factionsAppend at any point in the
running game, even after many other locality changes, without relying
on init lines.

Note that while this may seem like an unnecessary overhead for units
without per-unit loadouts (which it is), the overhead is very tiny
and happens only once (when transferring details about the unit), not
regularly for the same unit during gameplay.

Signed-off-by: freghar <freghar@dummy.tld>