Revert "remove Arsenal_Respawn compatibility in Custom_Factions"
The main problem of the unscheduled init EH is that Arma engine gear
loading is split into "backpack" and "everything else" with the init EH
being *between* the two.

After a weekend of trying to come up with workarounds revolving around
either

- moving the init EH registration + initial run to postInit
  - cannot work for JIP, the old logic is still used

- Local EH, catching the unit transfer from server->client

no workarounds are possible if we want to retain JIP/respawn/..
compatibility while guaranteeing locality of the unit during loadout
script execution (needed for some loadout commands).

The old (now current) way has thus a race:

1) unit initializes on the server (always)
2) init EH runs, spawning a loadout script in background
3) engine adds a backpack to the unit
   - and we're hoping our script still didn't hit the removeBackpack
4) unit is transferred to the client

The loadout script needs to start running after the engine adds the
backpack but finish running *before* the transfer-to-client (if it
happens). This is especially challenging on JIP where there's only
a few ms in between.

Therefore do not use Custom_Factions for players if you can help it,
or any Playable units in general. It's safe to use it for other AI
units provided you either risk it or add custom logic to loadout
scripts, waiting for the engine-initialized backpack to appear, so
you can remove it.

Calling factionsExec on unit init line is always safe (assuming you
don't combine it with other per-class loadouts), so another way is
to load loadout for Playable units using init lines.

Also see
https://forums.bistudio.com/topic/190610-something-is-adding-standard-backpack-after-init-eh/

This reverts commit 0964053642dc1656f2782bd9a90e505e1acb62f7.