call CODEs in fn_factionsExec without a spawn
This reiforces the "raw" nature of factionsExec - don't change
the execution environment for the mission creator, it's their job
to spawn this function if the context needs it.

As this function is completely separate from Load/Append and doesn't
rely on any state variables, init EH, etc., this is safe to do.

Signed-off-by: freghar <freghar@dummy.tld>